wip 10
As the previous commit, there was a class not found issue.
This commit solved it with proper reference name.

Additionally, i make an attempt to fix this random state of the elastic
search ppolicy ressource.

As minor change, i fixed the mistake with the java_home variable who was
used as a local.